issue: create new packages and restructure code
