+++ aliases.go
@@ -1,5 +1,7 @@
 package main
 
+import "github.com/ECL2022PAI01/splunk-service/handler"
+
 const UnhandleKeptnCloudEvent = "Unhandled Keptn Cloud Event : "
 
 // ServiceName specifies the current services name (e.g., used as source when sending CloudEvents)
@@ -8,5 +10,5 @@ const ServiceName = "splunk-service"
 // These variables facilitate tests
 var cloudEventListener = CloudEventListener
 var processKeptnCloudEvent = ProcessKeptnCloudEvent
-var handleConfigureMonitoringTriggeredEvent = HandleConfigureMonitoringTriggeredEvent
-var handleGetSliTriggeredEvent = HandleGetSliTriggeredEvent
+var handleConfigureMonitoringTriggeredEvent = handler.HandleConfigureMonitoringTriggeredEvent
+var handleGetSliTriggeredEvent = handler.HandleGetSliTriggeredEvent
