+++ handler/get_sli_triggered_event_handler_test.go
@@ -1,4 +1,4 @@
-package main
+package handler
 
 import (
 	"context"
@@ -12,9 +12,8 @@ import (
 	"testing"
 	"time"
 
+	"github.com/ECL2022PAI01/splunk-service/pkg/utils"
 	"github.com/cloudevents/sdk-go/v2/event/datacodec"
-	"github.com/joho/godotenv"
-	keptnv1 "github.com/keptn/go-utils/pkg/lib"
 	"github.com/keptn/go-utils/pkg/lib/v0_2_0/fake"
 
 	cloudevents "github.com/cloudevents/sdk-go/v2" // make sure to use v2 cloudevents here
@@ -29,9 +28,9 @@ import (
 // and the default result you await from splunk
 // Indicators given in get-sli.triggered.json should match indicators in the given sli file
 const (
-	getSliTriggeredEventFile              = "test/events/get-sli.triggered.json"
-	configureMonitoringTriggeredEventFile = "test/events/monitoring.configure.json"
-	sliFilePath                           = "./test/data/podtatohead.sli.yaml"
+	getSliTriggeredEventFile              = "../test/events/get-sli.triggered.json"
+	configureMonitoringTriggeredEventFile = "../test/events/monitoring.configure.json"
+	sliFilePath                           = "../test/data/podtatohead.sli.yaml"
 	defaultSplunkTestResult               = 1250
 )
 
@@ -63,67 +62,6 @@ func initializeTestObjects(eventFileName string, resourceServiceUrl string) (*ke
 	return ddKeptn, incomingEvent, err
 }
 
-// Tests the HandleMonitoringTriggeredEvent
-func TestHandleConfigureMonitoringTriggeredEvent(t *testing.T) {
-
-	//Building a mock resource service server
-	resourceServiceServer, err := buildMockResourceServiceServer(sliFilePath)
-	if err != nil {
-		t.Errorf("Error reading sli file : %s", err.Error())
-		t.Fail()
-	}
-	defer resourceServiceServer.Close()
-
-	//Building a mock splunk server
-	splunkServer := builMockSplunkServer()
-	defer splunkServer.Close()
-
-	//setting splunk credentials
-	env.SplunkPort = strings.Split(splunkServer.URL, ":")[2]
-	env.SplunkHost = strings.Split(strings.Split(splunkServer.URL, ":")[1], "//")[1]
-	env.SplunkApiToken = "apiToken"
-
-	//Initializing test objects
-	ddKeptn, incomingEvent, err := initializeTestObjects(configureMonitoringTriggeredEventFile, resourceServiceServer.URL+"/api/resource-service")
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	if incomingEvent.Type() == keptnv1.ConfigureMonitoringEventType {
-		incomingEvent.SetType(keptnv2.GetTriggeredEventType(keptnv2.ConfigureMonitoringTaskName))
-	}
-
-	data := &keptnv2.ConfigureMonitoringTriggeredEventData{}
-	err = incomingEvent.DataAs(data)
-
-	if err != nil {
-		t.Fatal("Error getting keptn event data")
-	}
-
-	err = HandleConfigureMonitoringTriggeredEvent(ddKeptn, *incomingEvent, data)
-
-	if err != nil {
-		t.Fatalf("Error: %s", err.Error())
-	}
-
-	gotEvents := len(ddKeptn.EventSender.(*fake.EventSender).SentEvents)
-
-	// Verify that HandleGetSliTriggeredEvent has sent 2 cloudevents
-	if gotEvents != 2 {
-		t.Fatalf("Expected two events to be sent, but got %v", gotEvents)
-	}
-
-	// Verify that the first CE sent is a .started event
-	if keptnv2.GetStartedEventType(keptnv2.ConfigureMonitoringTaskName) != ddKeptn.EventSender.(*fake.EventSender).SentEvents[0].Type() {
-		t.Fatal("Expected a configure-monitoring.started event type")
-	}
-
-	// Verify that the second CE sent is a .finished event
-	if keptnv2.GetFinishedEventType(keptnv2.ConfigureMonitoringTaskName) != ddKeptn.EventSender.(*fake.EventSender).SentEvents[1].Type() {
-		t.Fatal("Expected a configure-monitoring.finished event type")
-	}
-}
-
 // Tests the HandleSpecificSli function
 func TestHandleSpecificSli(t *testing.T) {
 	indicatorName := "test"
@@ -133,11 +71,11 @@ func TestHandleSpecificSli(t *testing.T) {
 
 	//Building a mock splunk server returning default responses when getting  get and post requests
 
-	splunkServer := builMockSplunkServer()
+	splunkServer := utils.BuildMockSplunkServer(defaultSplunkTestResult)
 	defer splunkServer.Close()
 
 	//Retrieving the mock splunk server credentials
-	splunkCreds := &splunkCredentials{
+	splunkCreds := &utils.SplunkCredentials{
 		Host:  strings.Split(strings.Split(splunkServer.URL, ":")[1], "//")[1],
 		Port:  strings.Split(splunkServer.URL, ":")[2],
 		Token: "apiToken",
@@ -174,10 +112,11 @@ func TestHandleGetSliTriggered(t *testing.T) {
 	defer resourceServiceServer.Close()
 
 	//Building a mock splunk server
-	splunkServer := builMockSplunkServer()
+	splunkServer := utils.BuildMockSplunkServer(defaultSplunkTestResult)
 	defer splunkServer.Close()
 
 	//setting splunk credentials
+	env := utils.EnvConfig{}
 	env.SplunkPort = strings.Split(splunkServer.URL, ":")[2]
 	env.SplunkHost = strings.Split(strings.Split(splunkServer.URL, ":")[1], "//")[1]
 	env.SplunkApiToken = "apiToken"
@@ -195,7 +134,7 @@ func TestHandleGetSliTriggered(t *testing.T) {
 		t.Fatalf("Error while getting keptn event data : %s", err.Error())
 	}
 
-	err = HandleGetSliTriggeredEvent(ddKeptn, *incomingEvent, data)
+	err = HandleGetSliTriggeredEvent(ddKeptn, *incomingEvent, data, env)
 
 	if err != nil {
 		t.Fatalf("Error : %s", err.Error())
@@ -241,47 +180,6 @@ func TestHandleGetSliTriggered(t *testing.T) {
 	}
 }
 
-// Tests the getSplunkCredentials function
-func TestGetSplunkCredentials(t *testing.T) {
-
-	godotenv.Load(".env.local")
-	env.SplunkApiToken = os.Getenv("SPLUNK_API_TOKEN")
-	env.SplunkHost = os.Getenv("SPLUNK_HOST")
-	env.SplunkPort = os.Getenv("SPLUNK_PORT")
-
-	sp, err := getSplunkCredentials()
-
-	if err == nil {
-		t.Logf("Splunk credentials : %v", sp)
-		if sp.Host == "" || sp.Port == "" || sp.Token == "" {
-			t.Fatal("If Host, Port or token are empty. An error should be returned")
-		}
-	} else {
-		t.Logf("Received expected error : %s", err.Error())
-	}
-}
-
-// Build a mock splunk server returning default responses when getting  get and post requests
-func builMockSplunkServer() *httptest.Server {
-
-	jsonResponsePOST := `{
-		"sid": "10"
-	}`
-	jsonResponseGET := `{
-		"results":[{"theRequest":"` + fmt.Sprint(defaultSplunkTestResult) + `"}]
-	}`
-	splunkResponses := make([]map[string]interface{}, 2)
-	splunkResponses[0] = map[string]interface{}{
-		"POST": jsonResponsePOST,
-	}
-	splunkResponses[1] = map[string]interface{}{
-		"GET": jsonResponseGET,
-	}
-	splunkServer := splunktest.MutitpleMockRequest(splunkResponses, true)
-
-	return splunkServer
-}
-
 // Build a mock resource service server returning a response with the content of the sli file
 func buildMockResourceServiceServer(filePath string) (*httptest.Server, error) {
 
@@ -302,4 +200,3 @@ func buildMockResourceServiceServer(filePath string) (*httptest.Server, error) {
 
 	return resourceServiceServer, nil
 }
-
